To-Do List App with jQuery for task management
The user can add a new item to a list of items.
The user can cross out an item from the list of items.
The user can delete an item from the list of items.
The user can change the order of items in the list of items.
The use gets an message if the field is empty.